# .scratch/mozhou-mvp/contracts/openapi.yaml
{
  "openapi": "3.0.3",
  "info": {
    "title": "墨舟 MoZhou API",
    "version": "1.0.0",
    "description": "AI 小说写作平台 — 全站 19 组端点契约（UVSD Contract-Driven）"
  },
  "servers": [
    {
      "url": "/",
      "description": "同源部署"
    }
  ],
  "paths": {
...
    "/api/v1/search": {
      "post": {
        "summary": "书源搜索（容错降级）",
        "operationId": "search",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "results": {
                      "type": "array"
                    },
                    "degraded": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "query": {
                    "type": "string"
                  }
                },
                "required": [
                  "query"
                ]
              }
            }
          }
        }
      }
    },
...
    "/api/v1/websearch": {
      "post": {
        "summary": "联网搜索",
        "operationId": "websearch",
        "responses": {
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "results": {
                      "type": "array"
                    },
                    "degraded": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "query": {
                    "type": "string"
                  }
                },
                "required": [
                  "query"
                ]
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "responses": {
      "RateLimited": {
        "description": "Too Many Requests：触发限流；Retry-After 头为建议等待秒数。桶与阈值见 api-contract.md §25",
        "content": { "application/json": { "schema": { "type": "object" } } }
      }
    },
...
  }
}